organized file tree better
