+++ components/ui/shared/theme-toggle.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import { Moon, Sun } from "lucide-react";
 import { useTheme } from "next-themes";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/shared/button";
 
 // Simple theme toggle button that switches between light and dark modes
 export function SimpleThemeToggle() {
